Add files via upload
The program is utilized to check whether the main board has passed the PBA test or not. It scans the
serial numbers of the main boards and retrieves their corresponding status and results from the
database, indicating whether they have passed the PBA test. Additionally, it displays the serial numbers
along with their status, with each status represented by a different color for easy identification
The program is developed using Python, Tkinter, and the MySQL Connector